Opening/Closing Doors Programming & Animations
Currently still working on trying to program doors and keys to work and have different levels of access. I'm not very good with animations but I'm getting there. This isn't finished yet but I'm at a good stopping point for the night and I'll pick up the rest later tomorrow.